Extract out footer and move some files
